requirements.txt
Django
django-simple-history
Faker
Jinja2
tox
urllib3
virtualenv
python-decouple
python-json-logger
requests
setuptools
coverage
whitenoise
djangorestframework
pandas
django-taggit
markdown
psycopg2
xlrd
Pillow
easy-thumbnails
django-ckeditor-5
django-unfold
django-axes
django-ratelimit
django-ratelimit2
django-two-factor-auth
phonenumbers
django-crispy-forms
crispy-bootstrap4
django-unicorn
django-robots
django_cryptography
django-encrypted-model-fields
django-auditlog
django-health-check
django-utils-six
reportlab
uvicorn
psycopg2-binary